chore(vite): convert core package configs to TypeScript
Rename vite.config.js → vite.config.ts and vite.tests.setup.js → .ts
across all packages/. Apply TypeScript-safe patterns and add proper
type annotations to the renderer test setup file. Also removes
deprecated deps.inline and hot config options.

Ref: https://github.com/podman-desktop/podman-desktop/issues/12105
