+++ package.json
@@ -32,7 +32,7 @@
     "build:preload-docker-extension": "cd ./packages/preload-docker-extension && vite build",
     "build:preload-webview": "cd ./packages/preload-webview && vite build",
     "build:preload:types": "dts-cb -i \"packages/preload/tsconfig.json\" -o \"packages/preload/exposedInMainWorld.d.ts\" && dts-cb -i \"packages/preload-docker-extension/tsconfig.json\" -o \"packages/preload-docker-extension/exposedInDockerExtension.d.ts\" && dts-cb -i \"packages/preload-webview/tsconfig.json\" -o \"packages/preload-webview/exposedInWebview.d.ts\"",
-    "build:renderer": "cross-env NODE_OPTIONS=--max-old-space-size=4096 vite -c packages/renderer/vite.config.js build",
+    "build:renderer": "cross-env NODE_OPTIONS=--max-old-space-size=4096 vite -c packages/renderer/vite.config.ts build",
     "build:ui": "cd ./packages/ui && pnpm run build",
     "compile": "cross-env MODE=production pnpm run build && electron-builder build --config .electron-builder.config.cjs --dir --config.asar=false",
     "compile:next": "cross-env MODE=production pnpm run build && electron-builder build --publish always --config .electron-builder.config.cjs",
