+++ packages/ui/vite.config.ts
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (C) 2023-2024 Red Hat, Inc.
+ * Copyright (C) 2024-2025 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,20 +15,34 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
-/* eslint-env node */
-import { join } from 'path';
-import { defineConfig } from 'vite';
 
-const PACKAGE_ROOT = __dirname;
+/* eslint-env node */
+import { join } from 'node:path';
+import * as path from 'node:path';
+import { svelte } from '@sveltejs/vite-plugin-svelte';
+import { svelteTesting } from '@testing-library/svelte/vite';
+import { defineConfig } from 'vite';
+import { fileURLToPath } from 'node:url';
+import tailwindcss from '@tailwindcss/vite';
+
+let filename = fileURLToPath(import.meta.url);
+const PACKAGE_ROOT = path.dirname(filename);
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  mode: process.env.MODE,
+  mode: process.env['MODE'],
   root: PACKAGE_ROOT,
   resolve: {
     alias: {
       '/@/': join(PACKAGE_ROOT, 'src') + '/',
     },
+  },
+  plugins: [tailwindcss(), svelte({ configFile: '../../svelte.config.js' }), svelteTesting()],
+  test: {
+    include: ['src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
+    globals: true,
+    environment: 'jsdom',
+    alias: [{ find: '@testing-library/svelte', replacement: '@testing-library/svelte/svelte5' }],
   },
   base: '',
   server: {
@@ -40,13 +54,12 @@ export default defineConfig({
     sourcemap: true,
     outDir: 'dist',
     assetsDir: '.',
+    lib: {
+      entry: 'src/lib/index.ts',
+      formats: ['es'],
+    },
 
     emptyOutDir: true,
     reportCompressedSize: false,
-  },
-  test: {
-    environment: 'node',
-    include: ['src/**/*.{test,spec}.?(c|m)[jt]s?(x)'],
-    passWithNoTests: true,
   },
 });
